requirements.txt
ultralytics==8.0.114
ipython
opencv-python-headless
pillow
Flask
flask_cors
moviepy